Add PT Timah Tbk Logo
PT Timah Tbk is part of MIND ID which is mainly engaged in tin, coal and nickel mining. As of the end of 2021, this company has a Mining Business License (IUP) in Riau and Bangka Belitung covering an area of ​​45,009 hectares which contains tin reserves of 300,000 tons. To support its business activities, this company also has a representative office in Jakarta.

More: https://timah.com